requirements.txt
pillow
setuptools
typer
pillow-avif-plugin